fix(models, migrations): modify OOD to keep consistency between Go structs and db tables

+++ internal/migrations/000001_create_artwork_images_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE artwork_images (
     id SERIAL PRIMARY KEY,
-    artwork_id INT NOT NULL REFERENCES artworks(id) ON DELETE CASCADE,
+    artwork_id BIGINT NOT NULL REFERENCES artworks(id) ON DELETE CASCADE,
     image_url TEXT NOT NULL,
     is_primary BOOLEAN DEFAULT FALSE,
     caption TEXT,
