+++ internal/migrations/000001_create_artwork_tags_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE artwork_tags (
-    artwork_id INT NOT NULL REFERENCES artworks(id) ON DELETE CASCADE,
-    tag_id INT NOT NULL REFERENCES tags(id) ON DELETE CASCADE,
+    artwork_id BIGINT NOT NULL REFERENCES artworks(id) ON DELETE CASCADE,
+    tag_id BIGINT NOT NULL REFERENCES tags(id) ON DELETE CASCADE,
     PRIMARY KEY (artwork_id, tag_id)
 );
