+++ internal/migrations/000001_create_forum_post_tags_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE forum_post_tags (
-    post_id INT NOT NULL REFERENCES forum_posts(id) ON DELETE CASCADE,
-    tag_id INT NOT NULL REFERENCES tags(id) ON DELETE CASCADE,
+    post_id BIGINT NOT NULL REFERENCES forum_posts(id) ON DELETE CASCADE,
+    tag_id BIGINT NOT NULL REFERENCES tags(id) ON DELETE CASCADE,
     PRIMARY KEY (post_id, tag_id)
 );
